fix dark mode hint color and Sam's review
